feat: Enhance chat message and add user settings
This commit enhances the chat message by adding user settings. It also introduces a new type 'Settings' and a new function 'badgeTemplate' to create badges. The 'buildBadge' function is updated to use the new 'badgeTemplate' function.

The Kernel class is updated to add a style element to the head. The style.css file is updated to add new styles for tooltip and chat badge.

The tailwind.config.js file is updated to remove the unused 'colors' module.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>